address comments, add a new overload of joinStr
