+++ dbms/src/Flash/Coprocessor/DAGUtils.h
@@ -22,6 +22,7 @@
 #include <Storages/Transaction/TiDB.h>
 #include <Storages/Transaction/Types.h>
 #include <grpcpp/impl/codegen/status_code_enum.h>
+#include <tipb/executor.pb.h>
 #include <tipb/select.pb.h>
 
 #include <unordered_map>
@@ -39,6 +40,10 @@ bool isWindowFunctionExpr(const tipb::Expr & expr);
 const String & getFunctionName(const tipb::Expr & expr);
 const String & getAggFunctionName(const tipb::Expr & expr);
 const String & getWindowFunctionName(const tipb::Expr & expr);
+String getExchangeTypeName(const tipb::ExchangeType & tp);
+String getJoinTypeName(const tipb::JoinType & tp);
+String getFieldTypeName(Int32 tp);
+String getJoinExecTypeName(const tipb::JoinExecType & tp);
 bool isColumnExpr(const tipb::Expr & expr);
 String getColumnNameForColumnExpr(const tipb::Expr & expr, const std::vector<NameAndTypePair> & input_col);
 NameAndTypePair getColumnNameAndTypeForColumnExpr(const tipb::Expr & expr, const std::vector<NameAndTypePair> & input_col);
